requirements.txt
Django>=5.1.6,<6.0
djangorestframework>=3.14.0
django-pandas>=0.6.7
xero-python>=4.0.0
pandas>=2.0.0
pandas-gbq>=0.19.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
pytz>=2023.3
requests>=2.31.0
apscheduler>=3.10.0
psycopg2-binary>=2.9.0

